ingest/requirements.txt
psycopg[binary]>=3.2
requests>=2.32
python-dotenv>=1.0
yt-dlp>=2025.1.1
pypdf>=5.1
beautifulsoup4>=4.12
markdown-it-py>=3.0
